+++ server/db.py
@@ -1,0 +1,6 @@
+from flask_mongoengine import MongoEngine
+
+db = MongoEngine()
+
+def initialize_db(app):
+    db.init_app(app)
